requirements.txt
pipecat-ai[cartesia,openai,silero,deepgram]
fastapi
uvicorn
python-dotenv
loguru
google-generativeai~=0.8.3
twilio
boto3